Add serial number storage and methods
Let's utilise the existing stockbook json storage methods
and morph into that of serial number storage.

Rationale: Update this serial number data whenever a new
stock is added to generate a unique serial number for the
stock and use this serial number as a key of calling other
methods.

Let's
- Create a new class AccQuantity that stores the accumulated
  amount of stocks for each source company,
- Create a new class SerialNumberSet that stores a pair of values,
  source and accQuantity,
- Update the respective LogicManager, StorageManager and
  ModelManager to incorporate the usage of the new
  SerialNumberSet,
- Duplicate and modify any existing codes that relate to
  StockBook under "storage" and "util" to store serial number
  data in a similar way to stockbook.json.