Modify implementation of add zoom link command
The current implementation of the add zoom link command requires users
to input the module name.

This can be further refined such that the command requires users to
input the module ID to aid efficiency.

Let's modify the addZoomLinkCommand class and all its dependent classes
to accept a module ID as the command argument.