Refactor Phone.java to Amount.java
Let's refactor Phone class to Amount class 
to be more reflective of a "amount" of an 
expense. 

Am amount comes in the format {dollars}.{cents}, 
although cents are optional.

This change includes updating fields in 
UI and tests.

TODO: decide if value of Amount should be stored
as dollars and cents instead (float accuracy).