Add working functionalities for normal delete
Add command support for delete item.

Basic case works, where all undeleted recipes connected to
this item, as a product or as a ingredient are also deleted

Case of adding a new item with same name as previous item is not
covered, as this would require a lot of amendments to AddItemCommand.