Adapt StringUtil to new Title field criteria
Updated the containsWordIgnoreCase method in StringUtil to allow
multiple words as keyword.

Since in PlaNus, the title attribute of a task is no longer required to
be a single word or a String without space compared to its previous
field in AddressBook3, the constrain for no multiple words as keyword
for checking title is removed.

Test cases adapted.