Refine testcases for a multitude of items.
Resolve an issue causing inconsistent testing outcomes. ie, at some
points, some tests would pass and then after rerunning would suddenly
fail.

Amended ItemBuilders to perform a deepcopy of items instead of a shallow
copy. This is due to the use of static test variables.

In future, all testcases should use static variables with respective
model component builders, as this gurantees that the variable is
deepcopied and will not change the base state of the eg APPLE