Redesign StudentRecord to be immutable
JavaFX ObservableList notifys the GUI to update when items are added or
deleted, but not when items are edited.

This results in the GUI not updating when a student's attendance is
marked (and their respective student record changed).

Let's make StudentRecord immutable so that whenever a student's record
is updated, the GUI will update correspondingly.