Simplify Year constructor
ParserUtil parse methods should only check validity and call the constructor. I have simplified parseYear so it it does so and does not break the abstraction barrier.
The Year regex already enforces a valid school type, so I have deleted the parseSchoolType method and added an assertion to Year constructor instead.
This also removes the dependency between the Storage and Model modules and  Logic since they no longer need the parseYear() method.