CategoryNotFoundException and Expense Equality
Update codebase to use CategoryNotFoundException as a runtime exception.
Handle category not found exceptions separately at a command level by throwing CommandException.
Add checks for CategoryNotFoundException in ExpenseBook and JsonSerializableExpenseBook.


Tighten notion of Expense equality to include BOTH date and amount. 

Update test and data files to accommodate stricter checks of having to add a category before being able to add an expense to that category.